Added a new file structure for the test files.
